requirements.txt
pillow
numpy
distance
